split function logic into files
